fix the compilation errors
